Fix some compiler warnings (implicit float to double conversion)

+++ src/ccstruct/ratngs.h
@@ -59,8 +59,8 @@ public:
     unichar_id_ = UNICHAR_SPACE;
     fontinfo_id_ = -1;
     fontinfo_id2_ = -1;
-    rating_ = 10.0;
-    certainty_ = -1.0;
+    rating_ = 10.0f;
+    certainty_ = -1.0f;
     script_id_ = -1;
     min_xheight_ = 0.0f;
     max_xheight_ = 0.0f;
@@ -170,13 +170,17 @@ public:
   bool PosAndSizeAgree(const BLOB_CHOICE &other, float x_height, bool debug) const;
 
   void print(const UNICHARSET *unicharset) const {
-    tprintf("r%.2f c%.2f x[%g,%g]: %d %s", rating_, certainty_, min_xheight_, max_xheight_,
+    tprintf("r%.2f c%.2f x[%g,%g]: %d %s",
+            static_cast<double>(rating_),
+            static_cast<double>(certainty_),
+            static_cast<double>(min_xheight_),
+            static_cast<double>(max_xheight_),
             unichar_id_, (unicharset == nullptr) ? "" : unicharset->debug_str(unichar_id_).c_str());
   }
   void print_full() const {
     print(nullptr);
     tprintf(" script=%d, font1=%d, font2=%d, yshift=%g, classifier=%d\n", script_id_, fontinfo_id_,
-            fontinfo_id2_, yshift_, classifier_);
+            fontinfo_id2_, static_cast<double>(yshift_), classifier_);
   }
   // Sort function for sorting BLOB_CHOICEs in increasing order of rating.
   static int SortByRating(const void *p1, const void *p2) {
